requirements.txt
ieee754==0.11
pyftdi==0.56.0
pyserial==3.5
PyYAML==6.0.2
